requirements.txt
fastapi
uvicorn
httpx
python-dotenv
jinja2
pytest
pytest-asyncio
pytest-mock